+++ cloud-infra-core/src/s3/dto.rs
@@ -126,7 +126,7 @@ pub struct LifecycleRule {
     #[serde(rename = "Status")]
     pub(crate) status: String,
     #[serde(rename = "ExpirationInDays", skip_serializing_if = "Option::is_none")]
-    pub(crate) expiration_in_days: Option<u32>,
+    pub(crate) expiration_in_days: Option<u16>,
     #[serde(rename = "Prefix", skip_serializing_if = "Option::is_none")]
     pub(crate) prefix: Option<String>,
     #[serde(rename = "ObjectSizeGreaterThan", skip_serializing_if = "Option::is_none")]
@@ -134,14 +134,13 @@ pub struct LifecycleRule {
     #[serde(rename = "ObjectSizeLessThan", skip_serializing_if = "Option::is_none")]
     pub(crate) object_size_less_than: Option<u32>,
     #[serde(rename = "AbortIncompleteMultipartUpload", skip_serializing_if = "Option::is_none")]
-    pub(crate) abort_incomplete_multipart_upload: Option<u32>, // TODO days
+    pub(crate) abort_incomplete_multipart_upload: Option<u16>,
     #[serde(rename = "NoncurrentVersionExpiration", skip_serializing_if = "Option::is_none")]
-    pub(crate) non_current_version_expiration: Option<u32>, // TODO days
+    pub(crate) non_current_version_expiration: Option<u16>,
     #[serde(rename = "Transitions", skip_serializing_if = "Option::is_none")]
     pub(crate) transitions: Option<Vec<LifecycleRuleTransition>>,
     #[serde(rename = "NoncurrentVersionTransitions", skip_serializing_if = "Option::is_none")]
     pub(crate) non_current_version_transitions: Option<Vec<NonCurrentVersionTransition>>
-    
     // #[serde(rename = "ExpiredObjectDeleteMarker", skip_serializing_if = "Option::is_none")]
     // pub(crate) expire_object_delete_marker: Option<bool> => cannot be specified with ExpirationInDays, ExpirationDate, or TagFilters.
     // "ExpirationDate": String => check the regex
